requirements.txt
google-generativeai
youtube-transcript-api
python-telegram-bot
openpyxl
requests
dotenv
svglib
reportlab
